:recycle: [Refactor] eventFormSection 분리 #easy
